+++ src/game/state.ts
@@ -1,10 +1,12 @@
+import { produce } from "immer";
+
 import { moveBlocks } from "./actions";
 import { Block, LevelState } from "./types";
 
 export const canPlaceAmount = (
   level: LevelState,
   columnIndex: number,
-  blocks: Block[]
+  blocks: Block[],
 ): number => {
   const column = level.columns[columnIndex];
   const spaceLeft = column.columnSize - column.blocks.length;
@@ -25,7 +27,7 @@ export const hasWon = (level: LevelState): boolean =>
       (col.type === "placement" &&
         col.columnSize === col.blocks.length &&
         col.blocks.every((b) => b.color === col.blocks[0].color)) ||
-      col.blocks.length === 0
+      col.blocks.length === 0,
   );
 
 export const isStuck = (level: LevelState): boolean => {
@@ -38,7 +40,7 @@ export const isStuck = (level: LevelState): boolean => {
   const countHidden = (level: LevelState) =>
     level.columns.reduce(
       (r, c) => r + c.blocks.filter((b) => b.revealed === true).length,
-      0
+      0,
     );
 
   const countCompleted = (level: LevelState) =>
@@ -46,7 +48,7 @@ export const isStuck = (level: LevelState): boolean => {
       (col) =>
         col.type === "placement" &&
         col.columnSize === col.blocks.length &&
-        col.blocks.every((b) => b.color === col.blocks[0].color)
+        col.blocks.every((b) => b.color === col.blocks[0].color),
     ).length;
 
   const topSignature = createSignature(level);
@@ -82,5 +84,35 @@ export const allShuffled = (level: LevelState): boolean =>
     (c) =>
       c.blocks.length < c.columnSize ||
       c.blocks.map((b) => b.color).filter((b, i, l) => l.indexOf(b) === i)
-        .length > 1
+        .length > 1,
   );
+
+export const getRevealedIndices = (
+  previousLevelState: LevelState,
+  newLevelState: LevelState,
+  columnIndex: number,
+) => {
+  const previous = previousLevelState.columns[columnIndex].blocks;
+  const size = newLevelState.columns[columnIndex].columnSize;
+  const prevBlockCount = previous.length;
+  const blockCount = newLevelState.columns[columnIndex].blocks.length;
+  const offset = prevBlockCount - blockCount;
+
+  return newLevelState.columns[columnIndex].blocks
+    .map((block, i) => ({
+      i: size - blockCount + i,
+      newlyRevealed: !!block.revealed && !previous[offset + i].revealed,
+    }))
+    .filter(({ newlyRevealed }) => newlyRevealed)
+    .map(({ i }) => i);
+};
+
+export const revealBlocks = (
+  levelState: LevelState,
+  revealed: { col: number; row: number }[],
+) =>
+  produce(levelState, (draft) => {
+    revealed.forEach(({ col, row }) => {
+      draft.columns[col].blocks[row].revealed = true;
+    });
+  });